\name{summary.TukeyRegion}
\alias{summary.TukeyRegion}
\title{Prints Summary of the Tukey Region}
\usage{
\method{summary}{TukeyRegion}(object, ...)
}
\description{
Prints basic information about the Tukey region.
}
\arguments{
\item{object}{
object of class \code{TukeyRegion} for which the summary should be printed.
}
\item{...}{
included for compatibility and should not be used.
}
}
\value{
Prints in the console basic information about the computed Tukey region.
}
\author{
Pavlo Mozharovskyi <pavlo.mozharovskyi@ensai.fr>
}
\seealso{
\code{\link{TukeyRegion}}, \code{\link{TukeyMedian}}
}
\examples{
# See examples in TukeyRegion or TukeyMedian
}
